New partition hash prefixes, remove prefix retentions and new range deletion

+++ opers/proc_part_mapper.go
@@ -6,7 +6,7 @@ func CalcProcessorPartitionMapping(mappingID string, partitionCount int,
 	processorCount int) map[int][]int {
 	ppm := map[int][]int{}
 	for partID := 0; partID < partitionCount; partID++ {
-		processorID := proc.CalcProcessorForPartition(mappingID, partID, processorCount)
+		processorID := proc.CalcProcessorForPartition(mappingID, uint64(partID), processorCount)
 		ppm[processorID] = append(ppm[processorID], partID)
 	}
 	return ppm
@@ -16,7 +16,7 @@ func CalcPartitionProcessorMapping(mappingID string, partitionCount int,
 	processorCount int) map[int]int {
 	ppm := map[int]int{}
 	for partID := 0; partID < partitionCount; partID++ {
-		processorID := proc.CalcProcessorForPartition(mappingID, partID, processorCount)
+		processorID := proc.CalcProcessorForPartition(mappingID, uint64(partID), processorCount)
 		ppm[partID] = processorID
 	}
 	return ppm
